requirements.txt
httpx
lark-oapi
google-genai
python-dotenv
